requirements.txt
streamlit
tensorflow
plotly
pandas
transformers
pillow
